backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==1.10.13
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
sqlalchemy==2.0.36
stripe==6.7.0
